frameworks/api-auth-examples/flask/requirements.txt
alembic==0.9.6
argcomplete==1.9.2
base58==0.2.4
bcrypt==3.1.4
boto3==1.4.8
botocore==1.8.1
certifi==2017.11.5
cffi==1.11.2
cfn-flip==0.2.5
chardet==3.0.4
click==6.7
docutils==0.14
durationpy==0.5
Flask==0.12.2
Flask-Bcrypt==0.7.1
Flask-Cors==3.0.3
Flask-Migrate==2.1.1
Flask-SQLAlchemy==2.3.2
future==0.16.0
gunicorn==19.7.1
hjson==3.0.1
idna==2.6
itsdangerous==0.24
Jinja2==2.10
jmespath==0.9.3
kappa==0.6.0
lambda-packages==0.19.0
Mako==1.0.7
MarkupSafe==1.0
placebo==0.8.1
psycopg2==2.7.3.2
py==1.5.2
pycparser==2.18
pyflakes==1.6.0
PyJWT==1.5.3
pytest==3.2.5
python-dateutil==2.6.1
python-editor==1.0.3
python-slugify==1.2.4
PyYAML==3.12
requests==2.18.4
s3transfer==0.1.11
six==1.11.0
SQLAlchemy==1.1.15
toml==0.9.3
tqdm==4.19.1
troposphere==2.1.0
Unidecode==0.4.21
urllib3==1.22
Werkzeug==0.12
wsgi-request-logger==0.4.6
WTForms==2.1
zappa==0.45.1
